+++ test/testdata/demo_long.py
@@ -201,8 +201,6 @@ CONST_NO_DOC = "SHOULD NOT APPEAR"
 class DataDemo:
     """
     This is an example for a dataclass.
-    Dataclasses generate a relatively pointless docstring by default,
-    but you can override it by providing your own (like here!).
 
     As usual, you can link to individual properties: `DataDemo.a`.
     """
@@ -213,6 +211,12 @@ class DataDemo:
     # This property has a type annotation but is not documented, so it does not show up.
     b: bool = field(repr=False, default=True)
     """This property is assigned to `dataclasses.field()`, which works just as well."""
+
+
+@dataclass
+class DataDemoExtended(DataDemo):
+    c: str = "42"
+    """A new attribute."""
 
 
 class EnumDemo(enum.Enum):